requirements.txt
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.5.7
botocore==1.8.21
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.8
django-modelcluster==3.1
django-storages==1.6.5
django-taggit==0.22.2
django-toolbelt==0.0.1
django-treebeard==4.2.0
djangorestframework==3.6.4
docutils==0.14
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
jmespath==0.9.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
static3==0.7.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.13.1
webencodings==0.5.1
Willow==1.0
